restructuring code + resolve comments for test.
